+++ src/function/arithmetic/sqrt.js
@@ -62,7 +62,10 @@ function factory (type, config, load, typed) {
    * @private
    */
   function _sqrtNumber (x) {
-    if (x >= 0 || config.predictable) {
+    if (isNaN(x)) {
+      return NaN
+    }
+    else if (x >= 0 || config.predictable) {
       return Math.sqrt(x)
     } else {
       return new type.Complex(x, 0).sqrt()