+++ test/function/statistics/median.test.js
@@ -67,6 +67,15 @@ describe('median', function () {
     ])), 3.5)
   })
 
+  it('should return NaN if any of the inputs contains NaN', function () {
+    assert.strictEqual(median([NaN]), NaN)
+    assert.strictEqual(median([1, NaN]), NaN)
+    assert.strictEqual(median([NaN, 1]), NaN)
+    assert.strictEqual(median([1, 3, NaN]), NaN)
+    assert.strictEqual(median([NaN, NaN, NaN]), NaN)
+    assert.strictEqual(median(NaN, NaN, NaN), NaN)
+  })
+
   it('should throw an error if called with invalid number of arguments', function () {
     assert.throws(function () { median() })
     assert.throws(function () { median([], 2, 3) })